secure tasks based on current user

+++ src/tasks/tasks.service.ts
@@ -14,8 +14,8 @@ export class TasksService {
 
     }
 
-    async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
+    async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+        return this.taskRepository.getTasks(filterDto, user);
     }
 
     // getAllTasks(): Task[] {
@@ -37,8 +37,8 @@ export class TasksService {
     //     return tasks;
     // }
 
-    async getTaskById(id: number): Promise<Task> {
-        const task = await this.taskRepository.findOne(id)
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const task = await this.taskRepository.findOne({where: {id, userId: user.id}})
         
         if (!task) {
             throw new NotFoundException();
@@ -51,15 +51,15 @@ export class TasksService {
         return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-       const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+       const result = await this.taskRepository.delete({id, userId: user.id}); 
        if (result.affected === 0) {
            throw new NotFoundException();
        }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
